// type.ts
import { FunctionComponent } from 'react'
import { IconType } from 'react-icons'

export interface IService {
    title: string;
    about: string;
    Icon: IconType; 
}

export interface ISkill {
    name: string;
    level: string;
    Icon: IconType;

}

export interface IProject {
    name: string;
    description: string;
    image_path: string;
    deployed_url: string;
    github_url: string;
    category: Category[];
    key_techs: string[];
}

export type Category = 'Front-end' | 'Data Analysis' | 'BI' | 'react' | 'python' | 'SQL' | 'JS' | 'sass' | 'Tableau' | 'MySQL' | 'PostgrSQL';